requirements.txt

aiogram==3.4.1
google-generativeai==0.3.2
python-dotenv==1.0.0
gspread==6.0.0
